The Renaissance
Massive change to the entire engine. Replaced the old project setup with a new approach. Removed the idea of multithreaded engine for standard single threaded approach. Resources are no longer all managed through a single interface, instead using the interface only to manage local files. Programmatically created resources will be able to interact with the rendering/audio interfaces as normal. Singleton interfaces (Math, Time, Resources etc.) that previously used statics to initialise the and provide access instead use a a single static in a Globals singleton that stores instances of and provides access to 'global' objects. This requires the use of a get function call to use these interfaces but cuts back on the number of static memory used. Unsure if this is better or worse needs to be tested. This project replaces work previously done, not up to implementing the  Rendering functionality again