Initial Rendering Implementation
Implemented initial form of the Rendering pipeline, providing an interface for modifications to the rendering process to be included and easily used. Currently a basic SceneRenderer has been included so that things (should) appear on screen soon. Need to implement a function that will allow for the flipping of backbuffers, but that is going to be future me's problem