Initial Audio Implementation
Initital implementation of the Audio system and its integration into the rest of the Engine. Provides functionality for playing SFX and Music as well as wrapping a number of the SDL_mixer functions to provide a uniform interface. Has built in mangement of channels (providing all interaction with SFX happens through the manager) and allows for the dynamic allocation of SFX Channels that can continue to a limit or beyond. System still needs testing