FString
Implemented a wrapper for a fixed CString, providing the functionality of a regular std::string without any of the memory allocation. Intended for use with fixed length values throughout the Engine (VirtualAxis names etc.) Initial tests show that it is working but have not done any fringe testing. Without including the declspec define on the FString class i recieve a number of compiler warnings and adding it give me a number of errors. Probably will need to figure that out soon. Stage is set to implement UI properly