Canvas Implementation and UIPanel
Created what (hopefully) will work mostly for the Canvas. UI elements can be created and are displayed but I have only implemented the UIPanel object currently. The ground work is there for interactable objects to work as well (Buttons, Textboxes etc.) but will need to implement those later. Testing needs to be done on what's there and loading UI elements from an Objx file needs to be setup