UILabel, UIButton and Dependencies Streamlining
Implemented the UILabel object for displaying text to the screen on the Canvas and the UIButton to allow for users to select actions via the UI system and perform actions. Fixed problems with the Canvas's casting and useage of the IUIAction and IUIBase objects allowing for correct useage of the data elements. Due to the number of shared libraries between the two projects i have combined the depeendencies folder to the solution directory to reduce the amount of duplicate files