Bomb Squad Map Generation Prototype
Implemented a possible method for generating procedural maps using cellular automota to create the basis for environements that can then be populated to suit the game. Current implementation works but the process used to find spawn points for the players to start at is incredably CPU intensive. Brute force solution employed to check all available points (multiple times depending on the number of players) to find the furthest point (on average) from the other spawn points. This is currently the bottle neck but the effects may not be as severe on a non-chromebook. Need to implement solution in C++