Finished Player Setup, Minor Fixes, Started Map Gen
Finished the Player Setup Scene, players can join with the specified control schemes and influence the animation that appears on screen before continuing to later setup stages. Fixed minor problems with teh SDL2_Engine as well as a large issue with the Virtual Axis monitoring that would cause axis' to influence the wrong values. Started porting the Map Generation code from JavaScript to C++ to allow for the procedural generation of maps to play the game on