Game Setup
Implemented all aspects of the Powerups for the game (Still untested) but theoretically can load all Objx files in the resources/Powerups folder location and create the icons for them. Created the property obejcts for the player so that the powerups that are picked up or gained throughout play can be stored somewhere easily accessible so that they can modify gameplay. Started creating the GameManager object but only so far as setup for a game. Need to implement game logic still. Removed the TestingGrounds project as BombSquad has overtaken it at this point